requirements.txt
aiogram==3.2.0
aiosqlite==0.19.0
python-dateutil==2.8.2
pytz==2023.3
aiofiles==23.2.1
SpeechRecognition==3.10.0
pydub==0.25.1
matplotlib==3.7.2
pandas==2.1.1
PyPDF2==3.0.1
python-docx==0.8.11
Pillow==10.0.1
python-dotenv==1.0.0
aiohttp==3.9.1
aiohttp-cors==0.7.0
psutil==5.9.6